+++ app/models/conversion/voluntary/tasks/handover.rb
@@ -1,4 +1,4 @@
-class Conversion::Voluntary::Tasks::Handover < TaskList::Task
+class Conversion::Voluntary::Tasks::Handover < TaskList::OptionalTask
   attribute :review
   attribute :notes
   attribute :meeting